+++ rel-script_test.go
@@ -23,7 +23,7 @@ add:
 `
 
 var bomIn = `
-Ref;Qnty;Value;Cmp name;Footprint;Description;Vendor;IPN;Datasheet
+Ref;Qty;Value;Cmp name;Footprint;Description;Vendor;IPN;Datasheet
 TP4, TP5;2;;Test point 2;;;;;
 R1, R2, ;2;;100K_100mw;;;;RES-006-0232;
 D1, D2, D13, D14;4;;diode;;;;DIO-023-0023;
@@ -31,7 +31,7 @@ D1, D2, D13, D14;4;;diode;;;;DIO-023-0023;
 `
 
 var bomExp = `
-Ref;Qnty;Value;Cmp name;Footprint;Description;Vendor;IPN;Datasheet
+Ref;Qty;Value;Cmp name;Footprint;Description;Vendor;IPN;Datasheet
 D1, D2, D14;3;;diode;;;;DIO-023-0023;
 R1, R2;2;;100K_100mw;;;;RES-006-0232;
 S3;1;;screw #4,2;;;;SCR-002-0002;